Refactor test.py into two client tests: unauthenticated and authenticated.
Move async test into test_async.py.
